% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fish.R
\docType{data}
\name{all_fish}
\alias{all_fish}
\title{all_fish}
\format{
character vector list
}
\source{
ESM 262

\describe {
 \item {no data title}{observed fish species} 
}
}
\usage{
all_fish
}
\description{
300 observations of fish species
}
\keyword{datasets}
