% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plastic_table.R
\docType{data}
\name{plastic_table}
\alias{plastic_table}
\title{plastic_table}
\format{
data frame with 4 variables and 19 rows
}
\source{
\describe {
  \item{carbon}{number of carbon atoms}
  \item{hydrogen}{number of hydrogen atoms}
  \item{oxygen}{number of oxygen atoms}
  \item{chlorine}{number of chlorine atoms}
}
}
\usage{
plastic_table
}
\description{
mock plastic sample data to be used as an example for the detect_plastic function
}
\keyword{datasets}
