% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slr_time.R
\name{slr_fun}
\alias{slr_fun}
\title{slr_time function}
\usage{
slr_fun(sea_level, elevation)
}
\arguments{
\item{sea_level}{}

\item{elevation}{}
}
\value{
year
}
\description{
Calculates time required for inundation given rate of sea level rise. 
Sea level rise time function (while loop ESM262 HW3). 
Returns year that a city will be inundated given a 3.6mm annual sea level rise (slr) rate (2021 global average). 
Takes starting local mean sea level (m) and city elevation (m) as inputs.
}
\examples{
slr_fun(sea_level = 1.1, elevation = 4.1)
[1] "The city will be flooded in year 2105"
}
